+++ PROJECT.py
@@ -59,7 +59,7 @@ pass    # place debug point here
     ]
 
     # HISTORY -----------------------------------------------
-    VERSION: tuple[int, int, int] = (0, 2, 12)
+    VERSION: tuple[int, int, int] = (0, 2, 13)
     TODO: list[str] = [
         "add TIMEOUT (use start in thread!) for print! use timeout for GETATTR!!!",
         [
@@ -78,8 +78,11 @@ pass    # place debug point here
         "..."
     ]
     NEWS: list[str] = [
-        "[OInfo] fix print",
-        "[TestPrimitives] move into pytest_aux",
+        ["[OInfo]",
+            "fix print repr (only if not equal to str!)",
+            "add mro in head!",
+            "use extemded ClsFullTypes with gens+iterables",
+         ]
     ]
 
     # FINALIZE -----------------------------------------------